hicore_api_extended/requirements.txt
# Core Dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.0

# Database
sqlalchemy>=1.4.0
alembic>=1.7.0

# Async
httpx>=0.23.0
aiofiles>=0.8.0

# Document Processing
PyPDF2>=2.0.0
python-docx>=0.8.11
beautifulsoup4>=4.10.0
requests>=2.26.0
python-multipart>=0.0.5

# Search and Web
ddg-search>=3.8.0
youtube-transcript-api>=0.6.1
duckduckgo-search>=3.8.0

# AI/ML (for Search-Analyzer)
sentence-transformers>=2.2.2
openai>=0.27.0
langchain-community>=0.0.10
faiss-cpu>=1.7.3  # or faiss-gpu if using GPU
numpy>=1.21.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0

# Utilities
python-jose[cryptography]>=3.3.0  # For JWT tokens
passlib[bcrypt]>=1.7.4  # For password hashing
python-multipart>=0.0.5  # For file uploads

# CORS
python-multipart>=0.0.5

# Environment
python-dotenv>=0.19.0
